perf(examples): parallelize watanabe DC-offset Optuna across worker processes
Extract simulation logic and constants into _oscillating_dc_helpers.py,
add _optimize_dc_worker.py subprocess worker, and rewrite the launch in
02_optimize_oscillating_dc.py to fan out N workers that all write to one
shared SQLite Optuna study. Avoids NEURON h singleton / macOS spawn
constraints; gallery-safe via conf.py ignore_pattern for the _ files.
